Properly implement LowerAllocObject
Summary:
Re-implement LowerAllocObject pass.
Previously, we implement LowerAllocObject by taking a list of users of the allocObjInst, assuming that the user list is sorted by dominance relationship and operate based on it.
This is not reliable, as optimization passes could change the order of user list, and user list provides no guarantee of dominance relationship.

This patch re-implements this pass in a proper way. We first walk down the dominance tree starting from the basic block of the allocObjInst, and collect basic blocks that contain users of the allocObjInst. Doing so ensures that we obtain these basic blocks in the order of dominance. To take care of the possible cases where we may have users of the allocObjInst in different branches, we ensure that in the basic block list we collected, every basic block must dominates the next, and whenever a branch is encountered, we stop appending more basic blocks (or even possibly remove basic blocks from the list).

Once we obtain these basic blocks, we then extract instructions (users) out of it, apply rules of the lowering to decide where to stop, and get the optimum strategy to lower it to achieve best bytecode size.

Reviewed By: avp

Differential Revision: D16364925

fbshipit-source-id: 46ed7d4516add56bfe2bd1e8151a8c4253f5c7a1