Add snapshot call at marker time
Summary:
In order to test heap snapshots on "real" heaps, add an option to synth benchmarks to take a heap
snapshot at a given marker point.
It's used by passing `-snapshot-at-marker $MARKER` to the synth benchmark runner.
The TraceInterpreter picks a file name based on the marker name and writes it to `/tmp`.

Example:
```
Wrote heap snapshot for marker "foo" to /tmp/foo-18a2c0.heapsnapshot
```

Reviewed By: willholen

Differential Revision: D16470019

fbshipit-source-id: d779134c42c4ba7ec630ee72686798b3670bd2ad