REPL: Load and write history to file (#49)
Summary:
Use libreadline to load and write command history to file. By default the history is written to `~/.hermes_history' and is created when does not exists.

I was not able to resolve the full path, so I had to check the home directory and append to `.hermes_history`.
Pull Request resolved: https://github.com/facebook/hermes/pull/49

Test Plan:
run the repl; run some commands; run the repl again.  verify history works.
look at the ~/.hermes_history file

Reviewed By: avp

Differential Revision: D16367773

Pulled By: mhorowitz

fbshipit-source-id: 48b053a1c95d7dcd29709860ff43aab8035b1344