Add option to force GC before printing stats
Summary:
Add option to force GC before printing stats when running a synth trace on the command-line. This allows measuring the live data when execution stopped at some interesting marker (under "heapInfo"->"Allocated bytes"). This enables estimating the impact of a diff on the live size of an application (rather than just total allocation, or allocated after the last organic collection).

Also add it for the hermes command-line tool, mostly for consistency.

Reviewed By: dulinriley

Differential Revision: D16603771

fbshipit-source-id: 69421e623a0cdc72e2071ee980fa100d0b2d6ded