Propagate x_facebook_sources field when merging source maps
Summary:
Preserves Facebook-specific source file metadata (encoded in the optional `x_facebook_sources` field) when it is found in an input source map. We do the minimal amount of work needed to pair each source filename ( = entry in the `sources` array) with its corresponding "metadata blob" ( = entry in the `x_facebook_sources`); the code is intentionally agnostic of what this blob represents or how it is encoded, beyond the fact that `null` represents missing metadata.

In order to have these opaque JSON values around for source map generation, we wrap them in instances of a new class, `JSONSharedValue`, which holds a shared reference to the underlying allocator of a JSON value, keeping it alive after the end of source map parsing.

Reviewed By: rubennorte

Differential Revision: D14951856

fbshipit-source-id: 952779cc0c2fd0fcd89e063485d25fb3ed67b592