Serializer and deserializer API.
Summary:
Serialization and Deserialization (S/D): This commit and the following commits
implements the first stage of S/D, where we are able to S/D th VM immediately
after global object initialization. The whole implementation are broken into
those commits for review purpose. The final commit will have tests that test
S/D end to end.

The basic idea is to S/D a graph of entities and each entity knows how to S/D
itself. Pointer values (GC pointers and native pointers) will need to be
relocated so they can point to the correct address after objects are
reconstructed during deserialization. During serialization, each address will
be givin a relocation id when it is first seen and this relocation id will be
written to the serialized stream. During deserialization each relocation id
will be able to know its new address after the entitie is reconstructed. Note
that we decide to do relocation after S/D an entity (Serializer::endObject and
Deserializer::endObject) instead of before so that it's easier to handle
VariableSizeRuntimeCell.

We distinguish between the following kinds of pointers: NormalPointers,
GCPointers, pointers in HermesValue, and function pointers (used by
NativeFunction and NativeConstructor). This commit handles the first three and
function pointers will be covered in later commits. We also map nullptr to id 0
in both Serializer and Deserializer so that we can deal with nullptr gracefully
during S/D.

This commit implements the serializer/deserializer API. The serialize stream
will have the following sections after finish serialization:

// ========================
// Serialized Data
// ========================
// charBuf_(used to deserilize static char* )
// ========================
// charBufSize(size of charBuf_ in bytes)
// ========================
// char16Buf_(used to deserilize static char16_t* )
// ========================
// char16BufSize(size of char16Buf_ in bytes)
// ========================
// mapSize

charBuf_ and char16Buf_ are used to store static char* and char16_t* string
literals and will be read and reconstructed first during deserialization. For
those string literals, the offset to the start of the buffer will be written to
the stream instead of a relocation id.

This commit also includes a simple unittest for the API.

Reviewed By: ridiculousfish

Differential Revision: D16262596

fbshipit-source-id: 616a825fcd29a3afc24b8d4fdc56f0b51c8ab369