Convert StringPrimitive snapshot to VTable function
Summary:
`StringPrimitive` is another type that is important to chrome heap snapshots, and can use
customization.
Hermes has two main different string types (internal and external strings), the latter don't take
any space in the JS heap, but take up `malloc` string space.
Report `ExternalStringPrimitive` as having an edge to a native node where the string data is kept.
The string data should be constant since strings are immutable, so it should persist across snapshots.

In order to keep the visualizer from choking on large input, large strings are truncated with a special
message, but the size remains the same.

Reviewed By: amnn

Differential Revision: D16528438

fbshipit-source-id: 5d5a483fe0bda57e9ab7bda8b6c759a05230adda