Check the integrity of the card boundary table.
Summary:
We're trying to debug a problem in GC, which shows up when we try find the first crossing object of a dirty card.

One possible explanation would be corruption by writes, by the mutator (or even outside of Hermes) between GCs.  This diff tries to detect that possibility: it takes a hash of the portion of the card boundary table used at the end of GCs, and ensures that this is the same at the start of the next GC.

Reviewed By: dulinriley

Differential Revision: D16769173

fbshipit-source-id: 68701afe7bf3a36bee1a3a8d90df217b54bfc8fc