Fix bug in extra debug code.
Summary:
In previous work, I added debug code to get more information a crash that occurs in production.  It turns out there was a subtle bug in the debugging code that was causing false positives that I didn't see in testing.  I quote from the explanation in the task below.  This diff fixes the problem explained there, so the debugging code will actually be useful.

The problem with the debugging code is actually pretty subtle, and I'll explain it here before fixing it.

OldGen::markYoungGenPointers loops over the segments in the old gen:

```
  auto segs = GCSegmentRange::concat(
      OldGenFilledSegmentRange::create(this),
      GCSegmentRange::singleton(&activeSegment()));

  size_t i = 0;
  while (AlignedHeapSegment *seg = segs->next()) {
```

Note that the last segment pointer yielded is to the active segment.  We're doing this to copy objects from the YG to the OG; this can cause a new segment to be materialized in the OG.  That moves the current active segment to the filled segment array (in the C++ sense of "move"), and sets the active segment to be a new segment.  In the case where we complain, this allocation of a new active segment happens *while* we're looking at dirty cards in the original active segment.  So the segment that seg points to represents a different range of heap addresses at the start of the iteration than it does at the end.  We immediately get a reference to the card table, though:

    auto &cardTable = seg->cardTable();

so that doesn't change.  This causes the code I added to fire with a false positive, because it first checks if the object found is in the "current segment":

      if (LLVM_UNLIKELY(!seg->contains(firstObj) || !firstObj->isValid())) {

when the contents of *seg, the active segment, change, this test fails, even though we're in the proper segment.  And the firstObj that's found is valid.

The fix is to save the bounds at the start, where we get the card table reference, and not refer to "seg" after the start of the loop iteration.

Reviewed By: tmikov

Differential Revision: D16820313

fbshipit-source-id: a0c2864d8c9d6bbd46895618f4a6fe22a082161f