Initial gradle setup
Summary:
Setting up Gradle for `fbjni` in preparation for the GitHub export.

This only covers the Java part for now; I'll add the CMake bits with the next diff.
The Java side already compiles, but without the dylib it will obviously fail at runtime as it is right now.

I tried to keep the structure as it is by bending the Gradle conventions. This avoids having to restructure all imports we have internally. If we feel the need to set up a sample app or something similar, we may want to reconsider this.

N.B. I'll ignore the shellcheck lints as this is just the output of `gradle init`.

Reviewed By: colriot, priteshrnandgaonkar

Differential Revision: D16915198

fbshipit-source-id: 654576cb53f7ca6c3b59551ddc59d547efc1ecf5