Parse StatementList in each case of a switch.
Summary:
The `parseSwitchStatement` function was looping through statements
and calling `parseStatement` repeatedly.

Unfortunately, StatementList in ES9.0 is not actually a list of Statements
any more - a StatementList can contain a Statement or a Declaration.
Thus, a `let` or `const` can be placed into the body of a `case`.
Hermes was failing to parse `let` and `const` because LexicalDeclarations
are not Statements.

Switching to `parseStatementList` requires allowing multiple
`until` Tokens to be provided to `parseStatementList` in this case.
In an attempt to allow flexibility in the long term, implement
`parseStatementList` as a templated function which can take any type
that can be passed to `checkN`.

Reviewed By: tmikov

Differential Revision: D16629198

fbshipit-source-id: 5ed8d447284b3cebd528e248c184c51da57cc11e