Finish Serialize/Deserialize Domain.
Summary:
With serialize/deserialize funtions implemented for RuntimeModule and
BCProvider in previous commits, we can have a complete implementation of
serialization/deserialization for Domain.

Runtime and Domain all maintain pointers to RuntimeModules, and we only want to
serialize/deserialize the contents of a certain RuntimeModule once. Therefore,
we need to decide on who is the "owner" and serialize/deserialize RuntimeModule
with its owner. We follow the same priciple of how memory are managed at
Runtime (as described in Design.md). That is: Domain owns a list of
RuntimeModules, and RuntimeModule owns a list of CodeBlocks.

Therefore, we call serialize/deserialize functions of RuntimeModules when we are
serializing Domains instead of when we are serializing Runtime fields, because
Domain "owns" the list of RumtimeModules instead of the Runtime. In other cases
when we see a pointer to a RuntimeModule or CodeBlock we simply do relocation
for the pointers.

With this commit, there is no need to handle specialCodeBlockDomain_ specially
for serialization/deserialization right after global object initialization.

Reviewed By: avp

Differential Revision: D16623986

fbshipit-source-id: 245a405d5da2429241a6f2272388d0d8f10d506f