Protect the Card Boundary table as often as possible to detect corrupting writes.
Summary:
Before this diff, we had inserted debugging code that "summarized" the card boundary tables of all old-gen segments at the end of GC, recording the portion that was summarized.  At the start of the next GC, we summarized the boundary tables again, making sure they hadn't changed.  We got an instance of the crash we're chasing in production with this instrumentation, and it indicated that the card boundary table had been corrupted.  But we don't know how.

This diff takes a next step: it uses memory protection on the boundary table in all times outside of legitimate modifications.  The hope is that whatever corrupting write is happening will be "caught in the act" by this.

Reviewed By: kodafb

Differential Revision: D17172877

fbshipit-source-id: 1292319f3f1be52af907d1df8e9da937deb0b724