Disable card boundary table protection.
Summary:
I realize today that there are too many things I don't understand about the crashes that have come in as a result of the card boundary table protection code.  I thought I understood the issues initially, but on further thought, I realized I had jumped to conclusions.  The proper, conservative, thing to do is to disable it for this branch cut, and try again next week, more carefully, making sure we can test on RN applications.

In addition, HaaS gets a dramatic (10-20X!) slowdown from this, we assume because of serialization of VM primitives in the kernel.  So we'll need to make it a GCConfig option, so we can turn it off for HaaS.

Reviewed By: dulinriley

Differential Revision: D17250711

fbshipit-source-id: 6cb52eb7b5372ddbc0c5006ae073e40690b1dcb9